# packages/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LangAI",
    "description": "Private dictionary representation training. Lexical knowledge is not conversational fluency.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/models/{model_id}/deploy": {
      "post": {
        "tags": [
          "Deployment & inference"
        ],
        "summary": "Deploy",
        "operationId": "deploy_models__model_id__deploy_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{model_id}/archive": {
      "post": {
        "tags": [
          "Deployment & inference"
        ],
        "summary": "Archive",
        "operationId": "archive_models__model_id__archive_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/embed": {
      "post": {
        "tags": [
          "Deployment & inference"
        ],
        "summary": "Embed",
        "operationId": "embed_v1_embed_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Embed"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/search": {
      "post": {
        "tags": [
          "Deployment & inference"
        ],
        "summary": "Search",
        "operationId": "search_v1_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Query"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/match": {
      "post": {
        "tags": [
          "Deployment & inference"
        ],
        "summary": "Match",
        "operationId": "match_v1_match_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Match"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/classify-language": {
      "post": {
        "tags": [
          "Deployment & inference"
        ],
        "summary": "Classify",
        "operationId": "classify_v1_classify_language_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Query"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/dictionary/{word}": {
      "get": {
        "tags": [
          "Deployment & inference"
        ],
        "summary": "Dictionary Word",
        "operationId": "dictionary_word_v1_dictionary__word__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "word",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Word"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Language"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictionaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{model_id}/playground": {
      "post": {
        "tags": [
          "Deployment & inference"
        ],
        "summary": "Playground",
        "operationId": "playground_models__model_id__playground_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Query"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClassificationResponse": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language"
          },
          "configured_language": {
            "type": "string",
            "title": "Configured Language"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "positive_probability": {
            "type": "number",
            "title": "Positive Probability"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "is_language": {
            "type": "boolean",
            "title": "Is Language"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          }
        },
        "type": "object",
        "required": [
          "language",
          "configured_language",
          "model_id",
          "positive_probability",
          "confidence",
          "is_language",
          "scope"
        ],
        "title": "ClassificationResponse"
      },
...
      "DictionaryEntrySchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "language_id": {
            "type": "string",
            "title": "Language Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "row_number": {
            "type": "integer",
            "title": "Row Number"
          },
          "original_row": {
            "additionalProperties": true,
            "type": "object",
            "title": "Original Row"
          },
          "headword": {
            "type": "string",
            "title": "Headword"
          },
          "normalized_headword": {
            "type": "string",
            "title": "Normalized Headword"
          },
          "definitions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Definitions"
          },
          "alternate_spellings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Alternate Spellings"
          },
          "part_of_speech": {
            "type": "string",
            "title": "Part Of Speech"
          },
          "dialect": {
            "type": "string",
            "title": "Dialect"
          },
          "examples": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Examples"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "approved": {
            "type": "boolean",
            "title": "Approved"
          },
          "training_eligible": {
            "type": "boolean",
            "title": "Training Eligible"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "id",
          "language_id",
          "source_id",
          "row_number",
          "original_row",
          "headword",
          "normalized_headword",
          "definitions",
          "alternate_spellings",
          "part_of_speech",
          "dialect",
          "examples",
          "notes",
          "source",
          "metadata",
          "approved",
          "training_eligible",
          "issues"
        ],
        "title": "DictionaryEntrySchema"
      },
      "DictionaryResponse": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/DictionaryEntrySchema"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "language",
          "entries"
        ],
        "title": "DictionaryResponse"
      },
      "Embed": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language"
          },
          "texts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 128,
            "minItems": 1,
            "title": "Texts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "language",
          "texts"
        ],
        "title": "Embed"
      },
      "EmbedResponse": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "dimensions": {
            "type": "integer",
            "title": "Dimensions"
          },
          "embeddings": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Embeddings"
          }
        },
        "type": "object",
        "required": [
          "language",
          "model_id",
          "dimensions",
          "embeddings"
        ],
        "title": "EmbedResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Match": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language"
          },
          "word": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Word"
          },
          "limit": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 5
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "language",
          "word"
        ],
        "title": "Match"
      },
      "MatchResponse": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "score_type": {
            "type": "string",
            "const": "cosine_similarity",
            "title": "Score Type"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/MatchResult"
            },
            "type": "array",
            "title": "Matches"
          }
        },
        "type": "object",
        "required": [
          "language",
          "model_id",
          "score_type",
          "matches"
        ],
        "title": "MatchResponse"
      },
      "MatchResult": {
        "properties": {
          "entry_id": {
            "type": "string",
            "title": "Entry Id"
          },
          "headword": {
            "type": "string",
            "title": "Headword"
          },
          "normalized_headword": {
            "type": "string",
            "title": "Normalized Headword"
          },
          "definitions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Definitions"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "dialect": {
            "type": "string",
            "title": "Dialect"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "definition": {
            "type": "string",
            "title": "Definition"
          }
        },
        "type": "object",
        "required": [
          "entry_id",
          "headword",
          "normalized_headword",
          "definitions",
          "source_id",
          "dialect",
          "score",
          "definition"
        ],
        "title": "MatchResult"
      },
...
      "Query": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language"
          },
          "query": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Query"
          },
          "limit": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 10
          },
          "mode": {
            "type": "string",
            "enum": [
              "semantic",
              "word-to-meaning",
              "meaning-to-word",
              "similar-words"
            ],
            "title": "Mode",
            "default": "semantic"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "language",
          "query"
        ],
        "title": "Query"
      },
...
      "SearchMatch": {
        "properties": {
          "entry_id": {
            "type": "string",
            "title": "Entry Id"
          },
          "headword": {
            "type": "string",
            "title": "Headword"
          },
          "normalized_headword": {
            "type": "string",
            "title": "Normalized Headword"
          },
          "definitions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Definitions"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "dialect": {
            "type": "string",
            "title": "Dialect"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "entry_id",
          "headword",
          "normalized_headword",
          "definitions",
          "source_id",
          "dialect",
          "score"
        ],
        "title": "SearchMatch"
      },
      "SearchResponse": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "score_type": {
            "type": "string",
            "const": "cosine_similarity",
            "title": "Score Type"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/SearchMatch"
            },
            "type": "array",
            "title": "Matches"
          }
        },
        "type": "object",
        "required": [
          "language",
          "model_id",
          "score_type",
          "matches"
        ],
        "title": "SearchResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}